Workshop

+++ src/main/java/com/apex/tech3/wallt_app/models/filters/WalletFilterOptions.java
@@ -2,11 +2,13 @@ package com.apex.tech3.wallt_app.models.filters;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class WalletFilterOptions {
     private Integer holderId;
-    private Long amountGreaterThan;
-    private Long amountLessThan;
+    private BigDecimal amountGreaterThan;
+    private BigDecimal amountLessThan;
     private Integer currencyId;
     private int page;
     private String sortBy;
